Create carrot cake recipe page
Create carrot cake page according to project
spesification. Each subpage should have a
main heading, an image, a paragraph
describing the recipe, a list of ingredients
and a list of steps. For my own convenience i
also added a link back to the index page at
the bottom of every recipe.